fix(precision): enforce FP16 dtype, enable TF32, align HPO batch range
P0: Added dtype=torch.float16 to autocast (was defaulting to BF16 on Blackwell)
P1: Added torch.set_float32_matmul_precision('high') to pipeline entrypoint
P2: Updated HPO batch_size search space from [64,128,256] to [256,512,1024]
Updated randd_log.md with full audit findings

+++ finrl_pro_ds/agents/deepscalper/bdq_agent.py
@@ -214,7 +214,7 @@ class DeepScalperBDQ:
         
         # Current Q-Values, Loss Computation under autocast
         # FIX H1: Use modern torch.amp.autocast (works on CPU + CUDA)
-        with torch.amp.autocast(device_type=self.device.type, enabled=self.use_amp):
+        with torch.amp.autocast(device_type=self.device.type, dtype=torch.float16, enabled=self.use_amp):
             q_dir, q_price, q_vol, _, pred_vol = self.policy_net(micro_state, private_state, macro_state)
         
             # Gather Q-values for taken actions
